--- main_1.py
@@ -1,9 +1,0 @@
-from screens_1 import main_screen_function
-
-
-def main() -> None:
-    main_screen_function()
-
-
-if __name__ == "__main__":
-    main()
