+++ mechanics.py
@@ -21,5 +21,4 @@ def load_image(name, colorkey=None):
         if colorkey == -1:
             colorkey = image.get_at((0, 0))
         image.set_colorkey(colorkey)
-
     return image
